+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [QATM-3183] Fix @dcos annotation if some container is null
 * [PIT-457] Step to make CCT installations independent from CCT version
 * [PIT-458] Step to make CCT uninstallations independent from CCT version
+* [QATM-3183] Try/catch added in @dcos to avoid errors if container is wrong
 
 ## 0.14.0 (March 09, 2020)